Tennis Pong
Basic two player Pong game made in the style of tennis. Done for my
second project in the first semester with the goal of making objects
that interact with each other. In order to play, one player moves using
the 'W' and 'S' keys while the other the 'up' and 'down' arrow keys. To
serve the ball press the either the 'D' or 'left' arrow key. Each score
is worth 15 points like in tennis. Starting from love as in tennis all
the way to 40 and the match point.